From cbeams:
Respect sticky/setuid bits when unzipping docs remotely

The spring-integration/docs/ directory on static.springframework.org
has sticky group-write permissions.  unzip is now invoked with the '-K'
flag to ensure that newly created directories respect these permissions.

This is important to ensure that different (ssh) users can release the
project without running into permissions errors.  Without -K, the unzip
command will preserve whatever permissions were present on the files and
directories at the time of archiving, and this usually means that
group-write is off.

Of course, all this assumes that users doing releases are part of the same
group - 'springorg' in this case.