Improve ROME tests
Additional testing revealed that the rometools version of ROME is not
compatible with Android. The Android port of the ROME feed reader was
reinstated in the previous commit. See the open issues against the
rometools project for more details.

https://github.com/rometools/rome/issues/155
https://github.com/rometools/rome/issues/163

This commit adds more complex tests for RSS and Atom feeds which verify
functionality that fails with the newer ROME version.

Issue: ANDROID-50