Add RestTemplate constructor for using custom message converters
This commits adds a new constructor for RestTemplate with a list
of message converters as argument. With this new constructor,
default message converters are never initialized. The default
constructor has also been modified to initialize the default set of
converters, which is a change from 1.0.

This commit also deprecates two constructors that accepted a boolean
parameter, which specified whether to initialize default message
converters or not.

These changes increase parity with RestTemplate in Spring Framework.

Issue: ANDROID-159