Disable connection reuse in SimpleClientHttpRequestFactory
Several versions of Android have issues with connection reuse and cause
errors such as an EOFException in certain cases. This commit sets the
"http.keepAlive" system property to "false" and also sets the HTTP
"connection" header with a value of "close". This is now the default
configuration for SimpleClientHttpRequestFactory. Connection reuse may
be enabled by setting "reuseConnection" property to true. HttpClient 4.3
is also now available as an alternative.

Issue: ANDROID-102, GITHUB-9